Update ObjectGraph, Edge, Node classes, add tests
The ObjectGraph class is updated to behave primarily as an Array, which
will be expected to contain only Node objects.  Initialization is
overridden to allow any number of Nodes or lists of Nodes to be passed
to the constructor.  The #+ and #push/#<< methods are overridden to
implement behavior more consistent with how the graph is expected to
handle Nodes.

The Node class has been introduced to store data about any given vertex
on the graph (specifically, resolved and unresolved edges and
metainformation to be used in edge resolution).  Intersecting nodes can
be summed to produce a new Node containing the the aggregated data from
them.  Non-intersecting nodes will produce nil when summed.

The Edge class is updated to only track the destination (directionality
of the edges is achieved by having the edge be owned by the origin node)
and the edge's weight.  Adding two edges produces a new edge with their
weights summed if the edges share their destination; otherwise nil is
returned.